moved tests to resp locations
